Refactor sample app navigation and modularize screen components.
* Extract `HomeScreen`, `LoginScreen`, and `BiometricKey` into dedicated files within the `cmp-sample-shared` module to improve project structure.
* Update `SampleAppNavigation` to handle consolidated `PasscodeResult` events, including verified, created, changed, and forgotten states.
* Ensure consistent backstack management by using `popUpTo(0)` when navigating to top-level screens like `HomeScreen` and `LoginScreen`.
* Update `PasscodeManager` to fully reset the internal state, including `isChangeFlow` and `isExternalAuthEnabled` flags, when clearing all security data.
* Remove the redundant `SettingsPasscodeScreen` route.
* Enhance `HomeScreen` with logic for toggling biometric authentication and triggering the passcode change flow.
* Refactor `BiometricKey` to handle biometric authentication results and provide appropriate user feedback via callbacks.
