+++ cmp-sample-shared/src/commonMain/kotlin/cmp/sample/shared/navigation/Route.kt
@@ -23,7 +23,4 @@ sealed class Route {
 
     @Serializable
     data object BiometricSetupScreen : Route()
-
-    @Serializable
-    data object SettingsPasscodeScreen : Route()
 }
